+++ src/containers/ToDo.js
@@ -1,4 +1,4 @@
-import React, { Component } from 'react';
+import React from 'react';
 import { toggleComplete } from '../actions';
 import { connect } from 'react-redux';
