+++ src/containers/ToDoList.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import ToDo from '../components/ToDo';
+import ToDo from './ToDo';
 import { connect } from 'react-redux';
 
 const ToDoList = ({todos}) => {
